requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.9.0
marshmallow==3.20.1
flask-cors==4.0.0
python-dateutil==2.8.2
email-validator==2.1.0
requests>=2.32.4